Fix --replace behavior regression
This fixes a regression where `--replace` was incorrectly considered a
targeted update. It removes a helper function indicating such, but the
incorrect logic is retained at the remaining usage site to avoid changes
in behavior.
